multi-agent-dev/requirements.txt
streamlit==1.28.0
langchain==0.0.292
langgraph==0.0.6
duckdb==0.8.1
anthropic==0.5.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.3
watchdog==3.0.0
python-dotenv==1.0.0